Refactoring: extract functions in 2024-11, use cache directly.
Function "walkIntoStone" was quite long and it was really easy to
extract code into separate functions to handle separate states, so it
was done.

There was no point in using a channel to make changes to the cache when
you could just do it directly. It was merely a holdover from an earlier
time when I thought I didn't need a mutex for the cache map. The thought
was that since concurrent reads aren't a problem, as long as I prevent
concurrent writes, it'll work without a mutex. I was wrong but since I
added the mutex, the channel also lost it's purpose.