2024-16: tune performance by not copying runner steps.
Previous steps for any given runner were keps in a slice. When runner
reached a spot with multiple options (go straight, turn left, or right),
the slice with the steps were copied. Otherwise the separate runners
would've possibly corrupted each others' steps.

The solution was to use a linked list instead. The first runner creates
the first item in the list and the added items point to the parent. Then
when multiple options are faced, the linked list is branched, and all
runners have a separate pointer to a leaf in the linked list tree. By
following the linked list, all runners only see their own steps and
eventually reach the first step that the first runner created.

The solution is very memory efficient and improved performance is due to
that. The linked list is only iterated and appended to. It's neither
searched nor modified. It's a perfect use case for a linked list tree.

Results with hyperfine follow. Almost twice as fast.

Before:
  Time (mean ± σ):   13.606 s ±  0.537 s  [User: 14.149 s, System: 0.320 s]
  Range (min … max): 12.835 s … 14.366 s  10 runs

After:
  Time (mean ± σ):    7.248 s ±  0.115 s  [User: 7.672 s, System: 0.103 s]
  Range (min … max):  7.104 s …  7.469 s  10 runs